NATIVE-386: Extract Pagination
